Split DBSP types and tests
